requirements.txt
annoy==1.16.2
dvc==0.94.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
glog==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
llvmlite==0.30.0
MarkupSafe==1.1.1
numba==0.46.0
numpy==1.17.4
openTSNE==0.3.11
pandas==0.25.3
psycopg2-binary==2.8.4
pyyaml==5.2
pynndescent==0.3.3
python-dateutil==2.8.1
python-gflags==3.1.2
pytz==2019.3
recsys-recoder==0.3.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
tqdm==4.40.2
Werkzeug==0.16.0
